requirements.txt
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
gunicorn==21.2.0
Flask-Limiter==3.13
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.10
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ordered-set==4.1.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
stripe==12.5.1
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
yarg==0.1.9
